file and mesia delete

+++ Modules/UploadFiles/Http/Controllers/MediaController.php
@@ -362,7 +362,10 @@ class MediaController extends Controller
         $file->delete();
         $media = media::whereId($request->mediaId)->first();
         $tempReturn = Lesson::find($request->lesson_id)->module('UploadFiles', 'media')->get();
-        $media->delete();
+
+        $mediaLessons = MediaLesson::where('media_id', $request->mediaId)->get();
+        if(count($mediaLessons) == 0)
+            $media->delete();
 
         if($media_type != null)
         {
